Move lib/client/SubscriptionConnector.js below lib/client/graphql/
