+++ lib/client/graphql/SubscriptionConnector.js
